+++ changelog.md
@@ -4,6 +4,8 @@
 
 - fetures:
   - qs-box
+  - added new prop `fieldsProps` to qs-form-generator
+  - added new prop `fieldProps` to qs-field
 
 ## 2.0.2 (10-28-2020)